+++ CHANGELOG.md
@@ -9,6 +9,7 @@ a Changelog](https://keepachangelog.com/en/1.0.0/) and the project adheres to
 Added:
 - Support for older browsers (back through IE 11, currently untested).
 - Dispatch return values in generator `yield` expressions.
+- `createAction.factory(...)` for fine-grained control in generators.
 
 Changed:
 - Enabled several compiler optimizations for smaller bundles.